Improve current system theme detection
Before it was checking the first 4 characters for "Dark". But version
manager messages in the beginning of the output started breaking that,
e.g.

> Using Node v16.20.2
> Dark

Now I strip out the newlines and then grab the last 4 characters to
check for "Dark".

It'd be ideal to avoid the version manager stuff altogether when
checking the system settings.